rename: Button 컴포넌트 파일명 index로 변경, 그에 따른 테스트 페이지 수정 #5
